Initial Commit of Rainbokeh
This is a simple web page that fades in one of my favorite background
images. It sits in a tab on one of my screens and is  used as a pleasant
background when I'm doing something else ont the other screen. Helps cut
down on distractions. This is always a pinned tab in my browser.

This could be modified easily to use any picture one might want to add.
It could also be modified to create a cool little slide show of background
images.

As far as I can tell this picture was taken by this person:
https://www.flickr.com/photos/romainguy/3908013145/